+++ unreal/AegisWar/Source/AegisWar/Public/WarInventoryWidget.h
@@ -6,12 +6,15 @@
 
 class SVerticalBox;
 class AWarPlayerState;
+class AWarCraftingStation;
 
 /** Native owner inventory view; selections carry the revision displayed to the user. */
 UCLASS()
 class AEGISWAR_API UWarInventoryWidget : public UUserWidget
 {
     GENERATED_BODY()
+public:
+    void SetCraftingStation(AWarCraftingStation* Station);
 protected:
     virtual TSharedRef<SWidget> RebuildWidget() override;
     virtual void NativeTick(const FGeometry& Geometry, float DeltaTime) override;
@@ -21,6 +24,8 @@ private:
     TSharedPtr<SVerticalBox> Rows;
     TWeakObjectPtr<AWarPlayerState> DisplayedState;
     int32 DisplayedRevision = INDEX_NONE;
+    TWeakObjectPtr<AWarCraftingStation> CraftingStation;
+    bool bStationSelected = false;
     FString Search;
     bool bSortByName = false;
 };